[minor / c] removes somes warning messages when compiling sources with clang, + uniformize mixed indent (spaces+tabs) to spaces only
git-svn-id: https://subversion.renater.fr/wimsdev/trunk@7076 b48fb02d-6135-0410-b1dd-deb767c81da9

+++ wims/src/Mathml/wims_mathml.cc
@@ -168,7 +168,7 @@ int main (int argc, char ** argv)
 	struct tm *tm;
 	gettimeofday(&tv, &tz);
 	tm=localtime(( const time_t * ) &tv.tv_sec);
-	snprintf(wims_mathml_id,sizeof(wims_mathml_id),"wims_mathml%ld",tv.tv_usec);
+	snprintf(wims_mathml_id,sizeof(wims_mathml_id),"wims_mathml%ld",long(tv.tv_usec));
 
 	if(bTexString){ // WIMS modification: reads a 'latex string' from commandline, if arg="--tex-string"
 	    char *input;
